+++ CHANGES.md
@@ -90,6 +90,7 @@ Changes in CUPS v2.5b1 (TBA)
 - Fixed encoding of `IPP_TAG_EXTENSION` values in IPP messages (Issue #913)
 - Fixed sending response headers to client (Issue #927)
 - Fixed `Host` header regression (Issue #967)
+- Fixed DNS-SD lookups of local services with Avahi (Issue #970)
 - Fixed CGI program initialization and validation of form checkbox and text
   fields.
 - Fixed finishing support in ippeveps.